securite try catch controller

+++ src/Controller/CategoryController.php
@@ -11,6 +11,7 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\Security\Core\Exception\AccessDeniedException;
 
 class CategoryController extends AbstractController
 {
@@ -20,12 +21,18 @@ class CategoryController extends AbstractController
      */
     public function showAllCategories(CategoryRepository $cateRepo, SubCategoryRepository $subCateRepo): Response
     {
-        $categories = $cateRepo->findAll(); //On lui demande d'afficher toutes les catégories
-        return $this->render('category/index.html.twig', [
-            'categories' => $categories,
-            'cate'=>$cateRepo->findAll(),
-            'SubCate'=>$subCateRepo->findAll()
-        ]);
+        try{
+            $this->denyAccessUnlessGranted('ROLE_ADMIN'); //je ne lui donne pas accès que si c'est un role admin
+            $categories = $cateRepo->findAll(); //On lui demande d'afficher toutes les catégories
+            return $this->render('category/index.html.twig', [
+                'categories' => $categories,
+                'cate'=>$cateRepo->findAll(),
+                'SubCate'=>$subCateRepo->findAll(),
+            ]);
+        }
+        catch(AccessDeniedException $ex){  //accès refusé car pas administrateur
+            return $this->redirectToRoute('home'); //retourne la page home si c'est pas l'admin
+        }
     }
 
      //AJOUTER/MODIFIER CATEGORIE
@@ -34,8 +41,9 @@ class CategoryController extends AbstractController
      * @Route("/category/add/{id}", name="add_cate")
      */
     public function updateOrAddCategory(Category $category=NULL, Request $request, EntityManagerInterface $em, CategoryRepository $cateRepo, SubCategoryRepository $subCateRepo){
+        try {
+            $this->denyAccessUnlessGranted('ROLE_ADMIN');
         if (!$category) {
-            
             $category = new Category;
         }
             $formCate=$this->createForm(CateType::class,$category); //creation du formulaire
@@ -55,7 +63,10 @@ class CategoryController extends AbstractController
                 'cate'=>$cateRepo->findAll(),
                 'SubCate'=>$subCateRepo->findAll()
             ]);
-        
+        } 
+        catch (AccessDeniedException $ex) {
+           return $this->redirectToRoute('home');
+        }
     }
 
 
@@ -63,10 +74,19 @@ class CategoryController extends AbstractController
     /**
      * @Route("/category/delete/{id}", name="delete_cate")
      */
-    public function deleteCategory(Category $category, EntityManagerInterface $em):Response{
-        $em->remove($category);
-        $em->flush();
-        
-        return $this->redirectToRoute('cate');
+    public function deleteCategory(Category $category, EntityManagerInterface $em, CategoryRepository $cateRepo, SubCategoryRepository $subCateRepo):Response{
+        try {
+            $this->denyAccessUnlessGranted('ROLE_ADMIN');
+            $em->remove($category);
+            $em->flush();
+            
+            return $this->redirectToRoute('cate',[
+                'cate'=>$cateRepo->findAll(),
+                'SubCate'=>$subCateRepo->findAll()
+            ]);
+        }
+        catch(AccessDeniedException $ex){
+            return $this->redirectToRoute('home');
+        }
     }
 }
